recouvrement/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2
django-bootstrap5==23.3
django-cors-headers==4.1.0
djangorestframework==3.14.0
filelock==3.12.2
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
leaflet==0.0.3
psycopg2-binary==2.9.6
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==3.4.4
urllib3==2.0.3
zope.event==4.6
zope.interface==6.0
